+++ pyproject.toml
@@ -51,6 +51,7 @@ dependencies = [
     "pyproj>=3.7.1,<4",
     "netcdf4>=1.7.2,<2",
     "pyshp>=2.3.1,<3",
+    "llvmlite>=0.44.0",
 ]
 dynamic = ["version"]
